+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 
 ### 🛠 Bug fixes
 * Transaction `priority` is applied as expected
+* Sender is not funded in raw transactions
 
 ## v21.1.9 (2021-10-21)
 ### 🛠 Bug fixes